refactor: extract find_in_file cluster to QueryEngine (Phase 2)
Move find_in_file and 5 helpers from CppAnalyzer to QueryEngine:
- _matches_glob
- _filter_results_by_files
- _find_in_files_glob
- _find_in_file_exact (split into _resolve_file_path and _match_item_to_file helpers to fix complexity)
- _get_path_suggestions
- find_in_file

CppAnalyzer keeps 1-line delegation for backward compatibility.
All state access in QueryEngine goes through self.analyzer back-reference.

All 1435 tests pass.
