+++ mcp_server/cpp_analyzer.py
@@ -2075,187 +2075,8 @@ class CppAnalyzer:
         return self.call_graph_service.get_call_path(from_function, to_function, max_depth)
 
     def find_in_file(self, file_path: str, pattern: str) -> Dict[str, Any]:
-        """Search for symbols within a specific file or files matching a glob pattern."""
-        # Detect if file_path is a glob pattern
-        glob_chars = set("*?[]")
-        is_glob = any(c in file_path for c in glob_chars)
-
-        if is_glob:
-            return self._find_in_files_glob(file_path, pattern)
-        else:
-            return self._find_in_file_exact(file_path, pattern)
-
-    def _matches_glob(self, indexed_file: str, glob_pattern: str) -> bool:
-        """Check if an indexed file matches a glob pattern using multiple strategies."""
-        import fnmatch
-
-        if fnmatch.fnmatch(indexed_file, glob_pattern):
-            return True
-        if fnmatch.fnmatch(indexed_file, "**/" + glob_pattern):
-            return True
-        if self.project_root:
-            try:
-                rel_path = str(Path(indexed_file).relative_to(self.project_root))
-                return fnmatch.fnmatch(rel_path, glob_pattern)
-            except ValueError:
-                pass
-        return False
-
-    def _filter_results_by_files(
-        self, items: List[Dict[str, Any]], matched_files: set
-    ) -> List[Dict[str, Any]]:
-        """Filter search results to only include items from specified files."""
-        results = []
-        for item in items:
-            _item_loc = item.get("definition") or item.get("declaration") or {}
-            item_file = _item_loc.get("file") or item.get("file", "")
-            if item_file in matched_files:
-                results.append(item)
-        return results
-
-    def _find_in_files_glob(self, glob_pattern: str, symbol_pattern: str) -> Dict[str, Any]:
-        """Search for symbols in files matching a glob pattern.
-
-        Args:
-            glob_pattern: Glob pattern to match files (e.g., '**/tests/**/*.cpp')
-            symbol_pattern: Symbol search pattern
-
-        Returns:
-            Dict with results, matched_files, and message
-        """
-        matched_files = [f for f in self.file_index.keys() if self._matches_glob(f, glob_pattern)]
-
-        if not matched_files:
-            return {
-                "results": [],
-                "matched_files": [],
-                "suggestions": self._get_path_suggestions(glob_pattern),
-                "message": f"No files found matching glob pattern '{glob_pattern}'",
-            }
-
-        all_classes = self.search_classes(symbol_pattern, project_only=False)
-        all_functions = self.search_functions(symbol_pattern, project_only=False)
-        matched_files_set = set(matched_files)
-        results = self._filter_results_by_files(all_classes + all_functions, matched_files_set)
-
-        return {
-            "results": results,
-            "matched_files": sorted(matched_files),
-            "message": f"Found {len(results)} symbols in {len(matched_files)} files matching '{glob_pattern}'",
-        }
-
-    def _find_in_file_exact(self, file_path: str, pattern: str) -> Dict[str, Any]:
-        """Search for symbols in a specific file (exact or suffix match).
-
-        Args:
-            file_path: Path to file (absolute, relative, or filename)
-            pattern: Symbol search pattern
-
-        Returns:
-            Dict with results, matched_files, suggestions (if empty), and message
-        """
-        results = []
-
-        # Search in both class and function results
-        all_classes = self.search_classes(pattern, project_only=False)
-        all_functions = self.search_functions(pattern, project_only=False)
-
-        # Try to resolve file path
-        abs_file_path = None
-        matched_file = None
-
-        # First, try exact absolute path match
-        if Path(file_path).is_absolute():
-            abs_file_path = str(Path(file_path).resolve())
-        else:
-            # Try relative to project root
-            if self.project_root:
-                potential_path = Path(self.project_root) / file_path
-                if potential_path.exists():
-                    abs_file_path = str(potential_path.resolve())
-
-        # Filter by file path
-        for item in all_classes + all_functions:
-            # Extract file from nested location object (definition or declaration)
-            _item_loc = item.get("definition") or item.get("declaration") or {}
-            item_file = _item_loc.get("file") or item.get("file", "")
-            if not item_file:
-                continue
-
-            # Match by absolute path or suffix
-            item_abs = str(Path(item_file).resolve()) if item_file else ""
-
-            if abs_file_path and item_abs == abs_file_path:
-                results.append(item)
-                matched_file = item_file
-            elif item_file.endswith(file_path) or item_abs.endswith(file_path):
-                results.append(item)
-                matched_file = item_file
-
-        if results:
-            return {
-                "results": results,
-                "matched_files": [matched_file] if matched_file else [],
-                "message": f"Found {len(results)} symbols in '{file_path}'",
-            }
-        else:
-            # No results - provide suggestions
-            suggestions = self._get_path_suggestions(file_path)
-            return {
-                "results": [],
-                "matched_files": [],
-                "suggestions": suggestions,
-                "message": f"No file found matching '{file_path}'. See suggestions for similar paths.",
-            }
-
-    def _get_path_suggestions(self, partial_path: str, max_suggestions: int = 5) -> List[str]:
-        """Get suggestions for similar file paths based on partial input.
-
-        Args:
-            partial_path: Partial path or filename to match
-            max_suggestions: Maximum number of suggestions to return
-
-        Returns:
-            List of suggested file paths
-        """
-        suggestions = []
-        partial_lower = partial_path.lower()
-
-        # Extract just the filename/basename for matching
-        partial_basename = Path(partial_path).name.lower()
-
-        # Also extract any path components for better matching
-        path_parts = [p.lower() for p in Path(partial_path).parts if p]
-
-        for indexed_file in self.file_index.keys():
-            indexed_lower = indexed_file.lower()
-            indexed_basename = Path(indexed_file).name.lower()
-
-            score = 0
-
-            # Exact basename match (highest priority)
-            if indexed_basename == partial_basename:
-                score += 100
-
-            # Basename contains partial
-            elif partial_basename in indexed_basename:
-                score += 50
-
-            # Path contains partial string
-            elif partial_lower in indexed_lower:
-                score += 30
-
-            # Check if path components match
-            for part in path_parts:
-                if part in indexed_lower:
-                    score += 10
-
-            if score > 0:
-                suggestions.append((score, indexed_file))
-
-        # Sort by score (descending) and take top suggestions
-        suggestions.sort(key=lambda x: (-x[0], x[1]))
-        return [path for _, path in suggestions[:max_suggestions]]
+        """Search for symbols within a specific file or files matching a glob pattern (delegates to query_engine)."""
+        return self.query_engine.find_in_file(file_path, pattern)
 
     def _find_class_definition_files(
         self,
